functions/requirements.txt
firebase-functions>=0.1.0
firebase-admin>=6.0.0
google-cloud-firestore>=2.11.0
google-cloud-storage>=2.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
google-genai>=1.0.0
